image_forge/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database & Queue
redis==5.0.1
sqlalchemy==2.0.25
aiosqlite==0.19.0

# FLUX / Diffusion
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.1
sentencepiece>=0.1.99

# Image processing
Pillow>=10.2.0
numpy>=1.26.0

# Utilities
httpx>=0.26.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0

# Optional: for Apple Silicon
# mlx>=0.0.10
